+++ src/main/java/com/muud/collection/repository/CollectionRepository.java
@@ -1,12 +1,11 @@
 package com.muud.collection.repository;
 
-import com.muud.collection.entity.Collection;
+import com.muud.collection.domain.Collection;
 import com.muud.playlist.entity.PlayList;
 import com.muud.user.entity.User;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
